+++ src/components/modal/task/CreateModalTask.tsx
@@ -1,16 +1,15 @@
-import { useQueryClient } from '@tanstack/react-query';
 import ModalLayout from '@layouts/ModalLayout';
 import ModalPortal from '@components/modal/ModalPortal';
 import ModalButton from '@components/modal/ModalButton';
 import ModalTaskForm from '@components/modal/task/ModalTaskForm';
 import useToast from '@hooks/useToast';
-import { useCreateStatusTask, useReadStatusTasks, useUploadTaskFile } from '@hooks/query/useTaskQuery';
+import useTaskFile from '@hooks/useTaskFile';
+import { useCreateStatusTask, useReadStatusTasks } from '@hooks/query/useTaskQuery';
 
 import type { SubmitHandler } from 'react-hook-form';
 import type { Project } from '@/types/ProjectType';
-import type { Task, TaskForm } from '@/types/TaskType';
+import type { TaskForm } from '@/types/TaskType';
 import type { ProjectStatus } from '@/types/ProjectStatusType';
-import type { FileUploadFailureResult, FileUploadSuccessResult } from '@/types/FileType';
 
 type CreateModalTaskProps = {
   project: Project;
@@ -19,11 +18,10 @@ type CreateModalTaskProps = {
 
 export default function CreateModalTask({ project, onClose: handleClose }: CreateModalTaskProps) {
   const createTaskFormId = 'createTaskForm';
-  const { toastSuccess, toastError } = useToast();
+  const { toastError } = useToast();
+  const { taskFilesUpload } = useTaskFile(project.projectId);
   const { mutateAsync: createTaskInfoMutateAsync } = useCreateStatusTask(project.projectId);
-  const { mutateAsync: createTaskFileMutateAsync } = useUploadTaskFile(project.projectId);
   const { statusTaskList } = useReadStatusTasks(project.projectId);
-  const queryClient = useQueryClient();
 
   const getLastSortOrder = (statusId: ProjectStatus['statusId']) => {
     const statusTask = statusTaskList.find((statusTask) => statusTask.statusId === Number(statusId));
@@ -32,32 +30,6 @@ export default function CreateModalTask({ project, onClose: handleClose }: Creat
       throw Error('프로젝트 상태가 존재하지 않습니다.');
     }
     return statusTask.tasks.length + 1;
-  };
-
-  const taskFilesUpload = async (taskId: Task['taskId'], files: File[]) => {
-    const createFilePromises = files.map((file) =>
-      createTaskFileMutateAsync({ taskId, file }).then(
-        (): FileUploadSuccessResult => ({ status: 'fulfilled', file }),
-        (error): FileUploadFailureResult => ({ status: 'rejected', file, error }),
-      ),
-    );
-
-    const results = (await Promise.allSettled(createFilePromises)) as PromiseFulfilledResult<
-      FileUploadSuccessResult | FileUploadFailureResult
-    >[];
-    const queryKey = ['projects', project.projectId, 'tasks'];
-    queryClient.invalidateQueries({ queryKey });
-
-    const fulfilledFileList: FileUploadSuccessResult[] = [];
-    const rejectedFileList: FileUploadFailureResult[] = [];
-    results
-      .map((result) => result.value)
-      .forEach((result) => {
-        if (result.status === 'fulfilled') fulfilledFileList.push(result);
-        else rejectedFileList.push(result);
-      });
-
-    if (fulfilledFileList.length > 0) toastSuccess(`${fulfilledFileList.length}개의 파일 업로드에 성공했습니다.`);
   };
 
   const handleSubmit: SubmitHandler<TaskForm> = async (taskFormData) => {
@@ -71,6 +43,7 @@ export default function CreateModalTask({ project, onClose: handleClose }: Creat
     if (files.length > 0) await taskFilesUpload(taskInfo.taskId, files);
     handleClose();
   };
+
   return (
     <ModalPortal>
       <ModalLayout onClose={handleClose}>
