+++ src/components/modal/task/UpdateModalTask.tsx
@@ -15,6 +15,7 @@ import { TASK_SETTINGS } from '@constants/settings';
 import { TASK_VALIDATION_RULES } from '@constants/formValidationRules';
 import useAxios from '@hooks/useAxios';
 import useToast from '@hooks/useToast';
+import useTaskFile from '@hooks/useTaskFile';
 import { useReadStatuses } from '@hooks/query/useStatusQuery';
 import {
   useAddAssignee,
@@ -46,6 +47,7 @@ export default function UpdateModalTask({ project, taskId, onClose: handleClose 
 
   const [keyword, setKeyword] = useState('');
   const { toastInfo, toastWarn } = useToast();
+  const { isValidTaskFile, taskFilesUpload } = useTaskFile(projectId);
   const { data: userList = [], loading, clearData, fetchData } = useAxios(findUserByProject);
   const searchCallbackInfo: ProjectSearchCallback = useMemo(
     () => ({ type: 'PROJECT', searchCallback: fetchData }),
@@ -105,25 +107,17 @@ export default function UpdateModalTask({ project, taskId, onClose: handleClose 
     deleteAssigneeMutate(user.userId);
   };
 
-  // ToDo: 일정 파일 업로드 작업시 같이 작업할 것
-  const updateFiles = (newFiles: FileList) => {
+  const updateTaskFiles = async (newFiles: FileList) => {
     if (taskFileList.length + newFiles.length > TASK_SETTINGS.MAX_FILE_COUNT) {
       return toastWarn(`최대로 등록 가능한 파일수는 ${TASK_SETTINGS.MAX_FILE_COUNT}개입니다.`);
     }
-  };
 
-  // ToDo: 일정 파일 업로드 API 작업 후 추가할 것
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { files } = e.target;
-    if (!files || files.length === 0) return;
-    updateFiles(files);
-  };
-
-  // ToDo: 일정 파일 업로드 API 작업 후 추가할 것
-  const handleFileDrop = (e: React.DragEvent<HTMLElement>) => {
-    const { files } = e.dataTransfer;
-    if (!files || files.length === 0) return;
-    updateFiles(files);
+    const validTaskFiles: File[] = [];
+    for (let i = 0; i < newFiles.length; i++) {
+      const file = newFiles[i];
+      if (isValidTaskFile(file)) validTaskFiles.push(file);
+    }
+    await taskFilesUpload(taskId, validTaskFiles);
   };
 
   const handleFileDeleteClick = (fileId: string) => deleteTaskFileMutate(Number(fileId));
@@ -207,8 +201,7 @@ export default function UpdateModalTask({ project, taskId, onClose: handleClose 
             label="첨부파일"
             files={taskFileList}
             accept={TASK_SETTINGS.FILE_ACCEPT}
-            onFileChange={handleFileChange}
-            onFileDrop={handleFileDrop}
+            updateFiles={updateTaskFiles}
             onFileDeleteClick={handleFileDeleteClick}
           />
         )}
